requirements.txt
python-telegram-bot==21.6
groq>=1.1.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
